Create component to create project, write unit tests
