+++ src/components/ProjectScreen/CreateProject/CreateProjectActions.tsx
@@ -1,8 +1,8 @@
 import { Dispatch } from "react";
-import * as backend from "../../backend";
-import { Project, defaultProject } from "../../types/project";
-import { setCurrentProject, ProjectAction } from "../Project/ProjectActions";
-import history from "../../history";
+import * as backend from "../../../backend";
+import { Project, defaultProject } from "../../../types/project";
+import { setCurrentProject, ProjectAction } from "../../Project/ProjectActions";
+import history from "../../../history";
 
 export const IN_PROGRESS = "CREATE_PROJECT_IN_PROGRESS";
 export type IN_PROGRESS = typeof IN_PROGRESS;
